Move file attribute tests into separate file
